chore(state): save current state
